multi-agent-ai-platform/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
langchain>=0.3.20
langgraph>=0.2.70
crewai>=1.10.0
pyautogen>=0.2.35
chromadb>=1.1.0
openai>=2.20.0
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.1

